Replace buggy scrublog migration migrations by proper one
