+++ pycroft/model/alembic/versions/2025_11_02_2036-82f43cfa0f98_add_scrublog.py
@@ -1,0 +1,48 @@
+"""Add scrublog
+
+Revision ID: 82f43cfa0f98
+Revises: b64618e97415
+Create Date: 2025-11-02 20:36:13.610741+00:00
+
+"""
+
+from alembic import op
+import sqlalchemy as sa
+from sqlalchemy.dialects import postgresql
+
+# revision identifiers, used by Alembic.
+revision = "82f43cfa0f98"
+down_revision = "b64618e97415"
+branch_labels = None
+depends_on = None
+
+
+def upgrade():
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.create_table(
+        "scrub_log",
+        sa.Column(
+            "executed_at",
+            sa.types.DateTime(timezone=True),
+            server_default=sa.text("CURRENT_TIMESTAMP"),
+            nullable=False,
+        ),
+        sa.Column("scrubber", sa.String(length=50), nullable=False),
+        sa.Column("info", postgresql.JSONB(astext_type=sa.Text()), nullable=False),
+        sa.Column("id", sa.Integer(), nullable=False),
+        sa.PrimaryKeyConstraint("id"),
+    )
+    op.create_index(op.f('ix_scrub_log_executed_at'), 'scrub_log', ['executed_at'], unique=False)
+    op.create_index(op.f('ix_scrub_log_info'), 'scrub_log', ['info'], unique=False)
+    op.create_index(op.f('ix_scrub_log_scrubber'), 'scrub_log', ['scrubber'], unique=False)
+
+    # ### end Alembic commands ###
+
+
+def downgrade():
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.drop_index(op.f('ix_scrub_log_scrubber'), table_name='scrub_log')
+    op.drop_index(op.f('ix_scrub_log_info'), table_name='scrub_log')
+    op.drop_index(op.f('ix_scrub_log_executed_at'), table_name='scrub_log')
+    op.drop_table('scrub_log')
+    # ### end Alembic commands ###
